대여기록 페이지 생성, 메인 페이지 페이지네이션 구현

+++ views/turn_view.py
@@ -1,7 +1,6 @@
-from flask import Blueprint, render_template, request, url_for, session, redirect, flash, g
+from flask import Blueprint, render_template, request, url_for, session, redirect, flash
 from models.models import *
 from db_connect import db
-import datetime
 
 bp = Blueprint('turn', __name__, url_prefix='/turn')
 
@@ -12,24 +11,37 @@ bp = Blueprint('turn', __name__, url_prefix='/turn')
 # 반납하기 버튼을 클릭할 경우 그 borrow기록에서 삭제하고 redirect를 한다.
 @bp.route('/')
 def home():
-    user_id = session['user_id']
-    borrows = BorrowInfo.query.filter(BorrowInfo.user_id == user_id).all()
-    book_list = []
-    
-    for borrow in borrows:
-        book = BookInfo.query.filter(BookInfo.id == borrow.book_id).first()
-        book_list.append(book)
-    
-    return render_template('turn.html', book_list=book_list, borrows = borrows)
+    if session.get('user_id'):
+        user_id = session['user_id']
+        borrows = BorrowInfo.query.filter(BorrowInfo.user_id == user_id).all()
+        book_list = []
+        
+        for i in range(len(borrows)):
+            book = BookInfo.query.filter(BookInfo.id == borrows[i].book_id).first()
+            book_list.append((book, borrows[i]))
+        
+        return render_template('turn.html', book_list=book_list)
+    else:
+        flash('로그인 후 사용해주세요.')
+        return redirect(url_for('user.login'))
 
 @bp.route('/back/<int:book_id>')
 def back(book_id):
-    borrow_info = BorrowInfo.query.filter(BorrowInfo.book_id == book_id).first()
-    book_info = BookInfo.query.filter(BookInfo.id == book_id).first()
-    book_info.stock += 1
-    db.session.delete(borrow_info)
-    db.session.commit()
-    
-    flash('정상적으로 반납 되었습니다.')
-    
-    return redirect(url_for('turn.home'))
+    if session.get('user_id'):
+        borrow_info = BorrowInfo.query.filter(BorrowInfo.book_id == book_id).first()
+        book_info = BookInfo.query.filter(BookInfo.id == book_id).first()
+        
+        #반납 테이블에 넣어주기
+        returninfo = ReturnInfo(user_id = session['user_id'], book_id = book_id, borrow_start = borrow_info.borrow_start, borrow_end = borrow_info.borrow_end)
+
+        book_info.stock += 1
+        db.session.delete(borrow_info)
+        db.session.add(returninfo)
+        db.session.commit()
+        
+        flash('정상적으로 반납 되었습니다.')
+        
+        return redirect(url_for('turn.home'))
+    else:
+        flash('로그인 후 사용해주세요.')
+        return redirect(url_for('user.login'))
